Add article list view.
