+++ backend/src/main/resources/templates/home.html
@@ -5,7 +5,7 @@
   <title>Title</title>
 </head>
 <body>
-  <p> HOla home</p>
-  <a href="article">Nuevo Articulo</a>
+  <p> home</p>
+  <a href="/article/new">New Article</a>
 </body>
 </html>
