+++ backend/src/main/resources/templates/new.html
@@ -8,7 +8,7 @@
   <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"
           integrity="sha256-k2WSCIexGzOj3Euiig+TlR8gA0EmPjuc79OEeY5L45g="
           crossorigin="anonymous"></script>
-  <script src="./js/init.js"></script>
+  <script src="/js/init.js"></script>
 
   <!-- etiqueta para crear estilos-->
   <style>
@@ -35,9 +35,9 @@
 </head>
 <body>
 
-<div class="text-center Semibold 36px bg-primary">
-  <h1>Nuevo Articulo</h1>
-</div>
+    <div class="text-center Semibold 36px bg-primary">
+      <h1>Nuevo Articulo</h1>
+    </div>
 
 
   <form class="col-sm-8">
@@ -121,9 +121,14 @@
 <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"
         integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa"
         crossorigin="anonymous"></script>
+    <div>
+      <a href="/home">home</a>
+    </div>
+    <div>
+      <a href="/article/list">Article List</a>
+    </div>
 
-<a href="home">home</a>
-<a href="list">listado de Articulos</a>
+
 
 </body>
